+++ orm/index.js
@@ -1,4 +1,5 @@
 module.exports = {
   baseModel: require('./base-model'),
-  baseRepository: require('./base-repository')
+  baseRepository: require('./base-repository'),
+  collection: require('./collection')
 };