+++ test/collection.non-standard.spec.js
@@ -1,0 +1,5 @@
+var dataLayer = require('./data-access/index.non-standard');
+
+describe('non-standard naming database', function() {
+  require('./tests/collection.js')(dataLayer);
+});