+++ test/collection.spec.js
@@ -1,0 +1,5 @@
+var dataLayer = require('./data-access/index');
+
+describe('standard naming database', function() {
+  require('./tests/collection.js')(dataLayer);
+});